removespikes: Use ds min/max thresholds if available
This changes the default logic to use the DS min/max thresholds unless binning
or threshold limits are specified.

To make analyze option useful in this mode that logic is slightly altered to proceed
further without writing any output before exiting.

Introduces dependency on Data::Dumper, but as it should be available on all
perl installations that shouldn't cause any trouble.